Added milvus vdb prototype impl
